+++ packages/storybook/src/spright/chat/input/chat-input-matrix.stories.ts
@@ -5,6 +5,11 @@ import {
     bodyFontColor
 } from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
 import { chatInputTag } from '@ni/spright-components/dist/esm/chat/input';
+import { buttonTag } from '@ni/nimble-components/dist/esm/button';
+import { selectTag } from '@ni/nimble-components/dist/esm/select';
+import { listOptionTag } from '@ni/nimble-components/dist/esm/list-option';
+import { iconAddTag } from '@ni/nimble-components/dist/esm/icons/add';
+import { chipTag } from '@ni/nimble-components/dist/esm/chip';
 import {
     sharedMatrixParameters,
     createMatrixThemeStory,
@@ -117,4 +122,29 @@ export const interactionsThemeMatrix: StoryFn = createMatrixThemeStory(
         active: [],
         focus: interactionStates
     })
+);
+
+export const slottedContent: StoryFn = createMatrixThemeStory(
+    html`
+        <p
+            style="
+            font: var(${bodyFont.cssCustomProperty});
+            color: var(${bodyFontColor.cssCustomProperty});
+            margin-bottom: 0px;
+            "
+        >
+            With attachments and footer-actions
+        </p>
+        <${chatInputTag} placeholder="Type a message">
+            <${chipTag} slot="attachments" removable>Placeholder.txt</${chipTag}>
+            <${buttonTag} slot="footer-actions" appearance="ghost" title="Attach" content-hidden>
+                <${iconAddTag} slot="start"></${iconAddTag}>
+                Attach
+            </${buttonTag}>
+            <${selectTag} appearance="underline" slot="footer-actions" title="Worker of the Week">
+                <${listOptionTag} value="option1">Inanimate Carbon Rod</${listOptionTag}>
+                <${listOptionTag} value="option2">Homer Simpson</${listOptionTag}>
+            </${selectTag}>
+        </${chatInputTag}>
+    `
 );